pyproject.toml
[tool.poetry]
name = "telegram_bot"
version = "0.1.0"
description = ""
authors = ["Fuck Poetry <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.9,<3.11"
aiogram = "^2.14.3"
python-dotenv = "^0.19.0"
vosk = "^0.3.30"
PySoundFile = "^0.9.0"
numpy = "^1.21.2"

[tool.poetry.dev-dependencies]
black = "^21.7b0"
isort = "^5.9.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
